_requirements.txt
pyautogui
requests
pillow
pystray